+++ services/datcore-adapter/src/simcore_service_datcore_adapter/api/rest/datasets.py
@@ -6,10 +6,13 @@ from fastapi import APIRouter, Depends, Header, Request
 from fastapi_pagination import Page, Params
 from fastapi_pagination.api import create_page, resolve_params
 from fastapi_pagination.bases import RawParams
+from models_library.api_schemas_datcore_adapter.datasets import (
+    DatasetMetaData,
+    FileMetaData,
+)
 from servicelib.fastapi.requests_decorators import cancel_on_disconnect
 from starlette import status
 
-from ...models.domains.datasets import DatasetsOut, FileMetaDataOut
 from ...modules.pennsieve import PennsieveApiClient
 from ..dependencies.pennsieve import get_pennsieve_api_client
 
@@ -26,7 +29,7 @@ _PENNSIEVE_CACHING_TTL_S: Final[int] = (
     "/datasets",
     summary="list datasets",
     status_code=status.HTTP_200_OK,
-    response_model=Page[DatasetsOut],
+    response_model=Page[DatasetMetaData],
 )
 @cancel_on_disconnect
 @cached(
@@ -39,7 +42,7 @@ async def list_datasets(
     x_datcore_api_secret: Annotated[str, Header(..., description="Datcore API Secret")],
     pennsieve_client: Annotated[PennsieveApiClient, Depends(get_pennsieve_api_client)],
     params: Annotated[Params, Depends()],
-) -> Page[DatasetsOut]:
+) -> Page[DatasetMetaData]:
     assert request  # nosec
     raw_params: RawParams = resolve_params(params).to_raw_params()
     assert raw_params.limit is not None  # nosec
@@ -57,7 +60,7 @@ async def list_datasets(
     "/datasets/{dataset_id}/files",
     summary="list top level files/folders in a dataset",
     status_code=status.HTTP_200_OK,
-    response_model=Page[FileMetaDataOut],
+    response_model=Page[FileMetaData],
 )
 @cancel_on_disconnect
 @cached(
@@ -71,7 +74,7 @@ async def list_dataset_top_level_files(
     x_datcore_api_secret: Annotated[str, Header(..., description="Datcore API Secret")],
     pennsieve_client: Annotated[PennsieveApiClient, Depends(get_pennsieve_api_client)],
     params: Annotated[Params, Depends()],
-) -> Page[FileMetaDataOut]:
+) -> Page[FileMetaData]:
     assert request  # nosec
     raw_params: RawParams = resolve_params(params).to_raw_params()
 
@@ -91,7 +94,7 @@ async def list_dataset_top_level_files(
     "/datasets/{dataset_id}/files/{collection_id}",
     summary="list top level files/folders in a collection in a dataset",
     status_code=status.HTTP_200_OK,
-    response_model=Page[FileMetaDataOut],
+    response_model=Page[FileMetaData],
 )
 @cancel_on_disconnect
 @cached(
@@ -106,7 +109,7 @@ async def list_dataset_collection_files(
     x_datcore_api_secret: Annotated[str, Header(..., description="Datcore API Secret")],
     pennsieve_client: Annotated[PennsieveApiClient, Depends(get_pennsieve_api_client)],
     params: Annotated[Params, Depends()],
-) -> Page[FileMetaDataOut]:
+) -> Page[FileMetaData]:
     assert request  # nosec
     raw_params: RawParams = resolve_params(params).to_raw_params()
     assert raw_params.limit is not None  # nosec
@@ -126,7 +129,7 @@ async def list_dataset_collection_files(
     "/datasets/{dataset_id}/files_legacy",
     summary="list all file meta data in dataset",
     status_code=status.HTTP_200_OK,
-    response_model=list[FileMetaDataOut],
+    response_model=list[FileMetaData],
 )
 @cancel_on_disconnect
 @cached(
@@ -139,7 +142,7 @@ async def list_dataset_files_legacy(
     x_datcore_api_key: Annotated[str, Header(..., description="Datcore API Key")],
     x_datcore_api_secret: Annotated[str, Header(..., description="Datcore API Secret")],
     pennsieve_client: Annotated[PennsieveApiClient, Depends(get_pennsieve_api_client)],
-) -> list[FileMetaDataOut]:
+) -> list[FileMetaData]:
     assert request  # nosec
     return await pennsieve_client.list_all_dataset_files(
         api_key=x_datcore_api_key,
